requirements.txt
chromadb>=0.4.0
pandas>=2.0.0
numpy>=1.24.0
sentence-transformers>=2.2.0
streamlit>=1.30.0
python-dotenv>=1.0.0
openai>=1.0.0
tavily-python>=0.5.0